+++ src/Siklid/Command/PostmanCommand.php
@@ -1,4 +1,5 @@
 <?php
+
 declare(strict_types=1);
 
 namespace App\Siklid\Command;
